src/web_app/requirements.txt
#devops
black==22.3.0
pytest==7.1.3
ruff == 0.1.4
#web
pandas == 2.0.3
pyspark == 3.5.0
matplotlib == 3.7.3
flask == 3.0.0
tensorflow == 2.15.0
seaborn == 0.13.0
transformers == 4.35.2
locust == 2.19.1
databricks-sql-connector==2.9.3
nltk == 3.8.1
wordcloud == 1.9.2
